python/safety_requirements_20736.txt
-i https://pypi.org/simple
anyio==4.12.1; python_version >= '3.9'
apscheduler==3.11.2; python_version >= '3.8'
certifi==2026.1.4; python_version >= '3.7'
charset-normalizer==3.4.2; python_version >= '3.7'
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
idna==3.11; python_version >= '3.8'
mysql-connector-python==9.4.0; python_version >= '3.9'
python-dotenv==1.1.1; python_version >= '3.9'
python-telegram-bot==22.6; python_version >= '3.10'
requests==2.32.4; python_version >= '3.8'
tzdata==2025.3; python_version >= '2'
tzlocal==5.3.1; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'